+++ database/migrations/2024_09_17_165452_create_pedidos_table.php
@@ -20,8 +20,8 @@ class CreatePedidosTable extends Migration
             $table->foreignId('restored_by')->nullable()->constrained('users');
             $table->foreignId('cliente_id')->constrained('clientes');
             $table->foreignId('endereco_id')->constrained('enderecos');
-            $table->double('total', 8 , 2);
-            $table->double('totalSemDesconto', 8 , 2);
+            $table->decimal('total', 8 , 2);
+            $table->decimal('totalSemDesconto', 8 , 2);
             $table->integer('porcentagem');
             $table->timestamp('restored_at')->nullable();
             $table->softDeletes();
